requirement.txt
flask
python-dotenv
psycopg2-binary
dotenv

